move ContentCard to _ypp directory
