+++ packages/atlas/src/components/_ypp/ContentCard/ContentCard.stories.tsx
