+++ packages/atlas/src/components/_ypp/ContentCard/ContentCard.tsx
@@ -1,12 +1,11 @@
 import { FC } from 'react'
 
+import { Pill } from '@/components/Pill'
+import { PillProps } from '@/components/Pill/types'
+import { Text } from '@/components/Text'
+import { ButtonProps, TextButton } from '@/components/_buttons/Button'
+import { SvgActionChevronR } from '@/components/_icons'
 import { useMediaMatch } from '@/hooks/useMediaMatch'
-
-import { Pill } from '../Pill'
-import { PillProps } from '../Pill/types'
-import { Text } from '../Text'
-import { ButtonProps, TextButton } from '../_buttons/Button'
-import { SvgActionChevronR } from '../_icons'
 
 export type ContentCardProps = {
   pill?: Exclude<PillProps, 'size' | 'variant'>
