+++ packages/atlas/src/components/_ypp/ContentCard/index.ts
