+++ packages/atlas/src/views/global/YppLandingView/YppLandingView.styles.ts
@@ -93,7 +93,6 @@ export const CardImageRow = styled(LayoutGrid)`
 `
 
 export const ImageContainer = styled.div<{
-  reverseOrderOnDesktop?: boolean
   positionOnMobile?: 'center' | 'unset' | 'flex-end'
 }>`
   position: relative;
